apic: Port driver for e2k
