+++ src/drivers/interrupt/Mybuild
@@ -1,4 +1,3 @@
-/*FIXME mybuild wrong packet name */
 package embox.driver.interrupt
 
 abstract module irqctrl_api { }
@@ -14,20 +13,6 @@ module at91_aic extends irqctrl_api {
 
 module i8259 extends irqctrl_api {
 	source "i8259.c", "i8259.h"
-}
-
-module ioapic extends irqctrl_api {
-	source "ioapic.c", "ioapic.h"
-
-	depends lapic
-}
-
-/* TODO: lapic doesn't depend on ioapic, only in current realization */
-module lapic {
-	source "lapic.c", "lapic.h"
-
-	//@NoRuntime
-	//depends ioapic
 }
 
 module irqmp extends irqctrl_api {
@@ -99,7 +84,6 @@ module xen extends irqctrl_api {
 	source "xen.c"
 	depends embox.arch.xen.traps
 }
-
 
 module integratorcp_pic extends irqctrl_api {
 	source "integratorcp_pic.c", "integratorcp_pic.h"
